COPY: expand the type of numcompleted to 64 bits
Integer overflow occurs without this when copied more than 2^31 rows,
under the `COPY ON SEGMENT` mode.

Errors happen when it is casted to uint64, the type of `processed` in
`CopyStateData`, third-party Postgres driver, which takes it as an
int64, fails out of range.

(back ported from commit 8d40268b2fae258769eb75c5ca9c7485b0cd1957)

+++ src/include/cdb/cdbsreh.h
@@ -99,7 +99,7 @@ extern CdbSreh *makeCdbSreh(bool is_keep, bool reusing_existing_errtable,
 extern void destroyCdbSreh(CdbSreh *cdbsreh);
 extern void HandleSingleRowError(CdbSreh *cdbsreh);
 extern void ReportSrehResults(CdbSreh *cdbsreh, int total_rejected);
-extern void SendNumRows(int numrejected, int numcompleted);
+extern void SendNumRows(int numrejected, int64 numcompleted);
 extern void SendNumRowsRejected(int numrejected);
 extern void ValidateErrorTableMetaData(Relation rel);
 extern bool IsErrorTable(Relation rel);
